Add files via upload
We were asked to create a folder named test, and inside that we were supposed to create a module, inside that we were supposed to write a logic to print all the even numbers. Inside the same folder we were supposed to create an ipynb file and call that module and take the input for the function from the user and also implement exception handling. Then we had to handle all the edge cases.